+++ lottery-core/src/test/java/de/alpharogroup/lottery/combinations/CombinationResolverTest.java
@@ -7,6 +7,7 @@ import org.testng.annotations.Test;
 
 import java.util.List;
 
+import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertTrue;
 
 /**
@@ -16,14 +17,42 @@ public class CombinationResolverTest
 {
 
 	/**
+	 * Test the method {@link CombinationResolver#getAllPossibleCombinationsCount(int, int)}
+	 */
+	@Test
+	public void testGgetAllPossibleCombinationsCount()
+	{
+		List<Integer> values;
+		long actual;
+		long expected;
+		// new scenario...
+		expected = 4;
+		values = ListFactory.newArrayList(1, 2, 3, 4);
+		actual = CombinationResolver.getAllPossibleCombinationsCount(values.size(), 3);
+		assertEquals(actual, expected);
+		// new scenario...
+		expected = 10;
+		values = ListFactory.newArrayList(1, 2, 3, 4, 5);
+		actual = CombinationResolver.getAllPossibleCombinationsCount(values.size(), 3);
+		assertEquals(actual, expected);
+		// new scenario...
+		expected = 7;
+		values = ListFactory.newArrayList(1, 2, 3, 4, 5, 6, 7);
+		actual = CombinationResolver.getAllPossibleCombinationsCount(values.size(), 6);
+		assertEquals(actual, expected);
+		// new scenario...
+		expected = 165;
+		values = ListFactory.newRangeList(1, 11);
+		actual = CombinationResolver.getAllPossibleCombinationsCount(values.size(), 3);
+		assertEquals(actual, expected);
+	}
+
+	/**
 	 * Test the method {@link CombinationResolver#getAllCombinations(List, int)}
 	 */
 	@Test
 	public void testGetAllCombinations()
 	{
-		Fraction fraction = new Fraction(49);
-		int intValue = fraction.intValue();
-		System.out.println(intValue);
 		List<Integer> values;
 		List<List<Integer>> actual;
 		List<List<Integer>> expected;
@@ -51,11 +80,6 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList(2, 4, 5));
 		expected.add(ListFactory.newArrayList(3, 4, 5));
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
-
-		expected = ListFactory.newArrayList();
-		values = ListFactory.newRangeList(1, 11);
-		actual = CombinationResolver.getAllCombinations(values, 3);
-		System.err.println(actual.size());
 		// new scenario...
 		expected = ListFactory.newArrayList();
 		values = ListFactory.newArrayList(1, 2, 3, 4, 5, 6, 7);
@@ -68,6 +92,10 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList(1, 3, 4, 5, 6, 7));
 		expected.add(ListFactory.newArrayList(2, 3, 4, 5, 6, 7));
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		values = ListFactory.newRangeList(1, 11);
+		actual = CombinationResolver.getAllCombinations(values, 3);
+		assertEquals(actual.size(), 165);
 	}
 
 	/**
@@ -90,6 +118,21 @@ public class CombinationResolverTest
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
 		// new scenario...
 		expected = ListFactory.newArrayList();
+		values = ListFactory.newArrayList("1", "2", "3", "4", "5");
+		actual = CombinationResolver.getCombinations(values, 3);
+		expected.add(ListFactory.newArrayList("1", "2", "3"));
+		expected.add(ListFactory.newArrayList("1", "2", "4"));
+		expected.add(ListFactory.newArrayList("1", "2", "5"));
+		expected.add(ListFactory.newArrayList("1", "3", "4"));
+		expected.add(ListFactory.newArrayList("1", "3", "5"));
+		expected.add(ListFactory.newArrayList("1", "4", "5"));
+		expected.add(ListFactory.newArrayList("2", "3", "4"));
+		expected.add(ListFactory.newArrayList("2", "3", "5"));
+		expected.add(ListFactory.newArrayList("2", "4", "5"));
+		expected.add(ListFactory.newArrayList("3", "4", "5"));
+		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		expected = ListFactory.newArrayList();
 		values = ListFactory.newArrayList("1", "2", "3", "4", "5", "6", "7");
 		actual = CombinationResolver.getCombinations(values, 6);
 		expected.add(ListFactory.newArrayList("1", "2", "3", "4", "5", "6"));
@@ -100,6 +143,10 @@ public class CombinationResolverTest
 		expected.add(ListFactory.newArrayList("1", "3", "4", "5", "6", "7"));
 		expected.add(ListFactory.newArrayList("2", "3", "4", "5", "6", "7"));
 		assertTrue(CollectionExtensions.isEqualCollection(actual, expected));
+		// new scenario...
+		values = ListFactory.newArrayList("1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11");
+		actual = CombinationResolver.getCombinations(values, 3);
+		assertEquals(actual.size(), 165);
 	}
 
 }